changed to parseFloat

+++ apps/frontend/public/scripts/main.js
@@ -10,10 +10,10 @@ function htmlToElement(html) {
 	return template.content.firstChild;
 }
 function paramJSON(paramName, defaultVal, minVal, maxVal, incrementVal) {
-	const parsedDef = parseInt(defaultVal);
-	const parsedMin = parseInt(minVal);
-	const parsedMax = parseInt(maxVal);
-	const parsedInc = parseInt(incrementVal);
+	const parsedDef = parseFloat(defaultVal);
+	const parsedMin = parseFloat(minVal);
+	const parsedMax = parseFloat(maxVal);
+	const parsedInc = parseFloat(incrementVal);
 	if(isNaN(parsedDef) || isNaN(parsedMin) || isNaN(parsedMax) || isNaN(parsedInc)) {
 		throw new TypeError();
 	}
